+++ src/main/java/com/YH/yeohaenghama/domain/itinerary/entity/Place.java
@@ -22,8 +22,8 @@ public class Place {
     private String placeNum;
     private String placeName;
     private String addr1;
-    private Integer mapx;
-    private Integer mapy;
+    private double mapx;
+    private double mapy;
     private String memo = null;
 
     @ManyToOne(fetch = FetchType.LAZY)
